lgu_back-main/lgu_back-main/requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework-simplejwt
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1